feat: Add JavaScript Basics Exercise 3 - Evaluate Excellence
Implemented Exercise 3 in the "JavaScript Basics" lesson with the function `solve`. The function takes a parameter `number` and evaluates if it is greater than or equal to 5.50. Depending on the condition, it prints 'Excellent' or 'Not excellent' to the console. This exercise enhances understanding of basic conditional statements and their application in JavaScript.