feat: Add JavaScript Basics Exercise 4 - Display Month by Number
Introduced Exercise 4 in the "JavaScript Basics" lesson with the function `solve`. This function takes a parameter `num` representing a month number and uses a `switch` statement to display the corresponding month's name. If the provided number does not match any valid month, it prints 'Error!' to the console. This exercise focuses on using switch statements for conditional branching in JavaScript.