feat: Add JavaScript Basics Exercise 8 - Calculate Circle Area
Introduced Exercise 8 in the "JavaScript Basics" lesson with the function `solve`. This function takes a parameter `input` and calculates the area of a circle if the input is a number. The result is printed with two decimal places using `toFixed(2)`. In case the input is not a number, it displays an informative message. This exercise emphasizes the usage of the `typeof` operator and conditional statements for handling different data types in JavaScript.