feat: Add string repetition function
Implemented a JavaScript function to repeat a given string a specified number of times. The function takes two parameters: the string to be repeated and the count indicating how many times it should be repeated. This exercise focuses on string manipulation and the use of loops in JavaScript. The result is logged to the console for verification.